Removed scripting, added epic and awesomesauce roblox importer, added support for custom and additional primitive meshes, file standard changes, UI changes, uhh, yeah. lots of stuff
Wowza its probably been at least a year since I pushed an update to this project. I've changed direction a bit. Now it's less of a game engine and more of a modeling tool. I mean, it was never really a game engine. I don't know what it was, but it's a modeling tool now. Its mainly designed to create large and explorable models as well as to introduce a new simplified model format which will most likely only be used by me. The intention is to preserve models by storing them in a simple JSON file which can be understood just by looking at it. If you still had a file for a 3D model 10 years in the future, would there still be software to load it? If not, you have to reverse engineer it, which is much harder to do for compressed and proprietary file formats. This would be easier if the purpose of each part of the file was obvious. Maybe you could write something to read the file since its so simple. Largely based on Roblox, models in this format can be made of primitive shapes but also allows for imported meshes, which are also stored in fairly obvious JSON. Of course, this takes up way more space, but whatever. You have to admit its kinda cool... also this very well could be my last update for this year. Happy 2026.